move adding tiles to end of action
added some debug "telemetry" for worst cycle since last action -
This showed ~20 msec, which convinced me that we should try to improve it.
Not surprisingly, adding the image to the tile was slow, so moving it helped.
Also structured the loop so we don't check a side that's been announced (i.e.
detected a touch) in the remaining time till lockout expires.
I now see ~5 msec worst case, more typical 2 - 3 msec for the worst cycle.